Done changes of RemoveLoggedInUser method and api

+++ src/app/common.service.ts
@@ -3,6 +3,7 @@ import { Subject, BehaviorSubject } from 'rxjs';
 import { ToastrService } from 'ngx-toastr';
 import { Router } from '@angular/router';
 import { BaseClass } from 'src/app/classes/BaseClass'
+import { AuthenticationService } from './services/authentication.service';
 
 @Injectable({
   providedIn: 'root'
@@ -11,7 +12,7 @@ export class CommonService {
   private baseClassObj = new BaseClass();  
   public authTokenstr:string = "The remote server returned an error: (401) Unauthorized.";  
 
-  constructor(private toastr: ToastrService, private router: Router) { }
+  constructor(private toastr: ToastrService, private router: Router,  private auth: AuthenticationService,) { }
   
   // Declaration
   private commonData = new Subject<any>();
@@ -32,11 +33,35 @@ export class CommonService {
     }
 }
 
-  public RemoveLoggedInUser(toastMsg){   
-     this.toastr.error('',toastMsg,this.baseClassObj.messageConfig); 
-        sessionStorage.clear();
-        localStorage.clear();
-        this.router.navigateByUrl('/login');              
+  public RemoveLoggedInUser(toastMsg){
+
+    let cmpName=window.localStorage.getItem("selectedComp");
+    let guid=window.localStorage.getItem("GUID");
+    let username=window.localStorage.getItem("loggedInUser");
+    let serviceURL = JSON.parse(window.localStorage.getItem('arrConfigData'));
+
+    this.auth.removeCurrentUser(serviceURL.service_url,cmpName,guid,username).subscribe(
+      data => {       
+        if(data == true){
+          if(toastMsg != ''){
+            this.toastr.error('',toastMsg,this.baseClassObj.messageConfig);
+          }
+          sessionStorage.clear();
+          localStorage.clear();
+          this.router.navigateByUrl('/login');
+        }
+        else{
+         // this.toastr.error('',toastMsg,this.baseClassObj.messageConfig); 
+         console.log("Cannot Log Out");
+        }
+      },
+      error => {
+          console.log(error);  
+          //this.toastr.error('',toastMsg,this.baseClassObj.messageConfig);              
+        } 
+    )
+
+                    
   }
